src/Backend/requirements.txt
flask
flask-mysqldb
flask-cors
python-dotenv
mysql-connector-python

